refactor(backend): split database/commands.rs by domain (phase 1)
Break 1270-line commands.rs (65 tauri commands) into domain
files under database/commands/:

  ai.rs (6), connections.rs (13), live_monitor.rs (2),
  mutation.rs (11), query.rs (9), query_builder.rs (2),
  schema.rs (2), schema_export.rs (2), scripts.rs (4),
  seeding.rs (1), settings.rs (4), snippet_folders.rs (4),
  snippets.rs (5)

No logic change. mod.rs re-exports every symbol so existing
database::commands::X call sites (lib.rs invoke_handler,
maintenance.rs helpers) resolve unchanged.

Part of docs/backend-refactor-plan.md.
